solana_client_traits

Trait SyncClient

pub trait SyncClient {

    // Required methods
    fn send_and_confirm_message<T: Signers + ?Sized>(
        &self,
        keypairs: &T,
        message: Message,
    ) -> Result<Signature>;
    fn send_and_confirm_instruction(
        &self,
        keypair: &Keypair,
        instruction: Instruction,
    ) -> Result<Signature>;
    fn transfer_and_confirm(
        &self,
        lamports: u64,
        keypair: &Keypair,
        pubkey: &Pubkey,
    ) -> Result<Signature>;
    fn get_account_data(&self, pubkey: &Pubkey) -> Result<Option<Vec<u8>>>;
    fn get_account(&self, pubkey: &Pubkey) -> Result<Option<Account>>;
    fn get_account_with_commitment(
        &self,
        pubkey: &Pubkey,
        commitment_config: CommitmentConfig,
    ) -> Result<Option<Account>>;
    fn get_balance(&self, pubkey: &Pubkey) -> Result<u64>;
    fn get_balance_with_commitment(
        &self,
        pubkey: &Pubkey,
        commitment_config: CommitmentConfig,
    ) -> Result<u64>;
    fn get_minimum_balance_for_rent_exemption(
        &self,
        data_len: usize,
    ) -> Result<u64>;
    fn get_signature_status(
        &self,
        signature: &Signature,
    ) -> Result<Option<TransactionResult<()>>>;
    fn get_signature_status_with_commitment(
        &self,
        signature: &Signature,
        commitment_config: CommitmentConfig,
    ) -> Result<Option<TransactionResult<()>>>;
    fn get_slot(&self) -> Result<u64>;
    fn get_slot_with_commitment(
        &self,
        commitment_config: CommitmentConfig,
    ) -> Result<u64>;
    fn get_transaction_count(&self) -> Result<u64>;
    fn get_transaction_count_with_commitment(
        &self,
        commitment_config: CommitmentConfig,
    ) -> Result<u64>;
    fn get_epoch_info(&self) -> Result<EpochInfo>;
    fn poll_for_signature_confirmation(
        &self,
        signature: &Signature,
        min_confirmed_blocks: usize,
    ) -> Result<usize>;
    fn poll_for_signature(&self, signature: &Signature) -> Result<()>;
    fn get_latest_blockhash(&self) -> Result<Hash>;
    fn get_latest_blockhash_with_commitment(
        &self,
        commitment_config: CommitmentConfig,
    ) -> Result<(Hash, u64)>;
    fn is_blockhash_valid(
        &self,
        blockhash: &Hash,
        commitment: CommitmentConfig,
    ) -> Result<bool>;
    fn get_fee_for_message(&self, message: &Message) -> Result<u64>;

}

Required Methods

fn send_and_confirm_message<T: Signers + ?Sized>( &self, keypairs: &T, message: Message, ) -> Result<Signature>

Create a transaction from the given message, and send it to the server, retrying as-needed.

fn send_and_confirm_instruction( &self, keypair: &Keypair, instruction: Instruction, ) -> Result<Signature>

Create a transaction from a single instruction that only requires a single signer. Then send it to the server, retrying as-needed.

fn transfer_and_confirm( &self, lamports: u64, keypair: &Keypair, pubkey: &Pubkey, ) -> Result<Signature>

Transfer lamports from keypair to pubkey, retrying until the transfer completes or produces and error.

fn get_account_data(&self, pubkey: &Pubkey) -> Result<Option<Vec<u8>>>

Get an account or None if not found.

fn get_account(&self, pubkey: &Pubkey) -> Result<Option<Account>>

Get an account or None if not found.

fn get_account_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig, ) -> Result<Option<Account>>

Get an account or None if not found. Uses explicit commitment configuration.

fn get_balance(&self, pubkey: &Pubkey) -> Result<u64>

Get account balance or 0 if not found.

fn get_balance_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig, ) -> Result<u64>

Get account balance or 0 if not found. Uses explicit commitment configuration.

fn get_minimum_balance_for_rent_exemption(&self, data_len: usize) -> Result<u64>

fn get_signature_status( &self, signature: &Signature, ) -> Result<Option<TransactionResult<()>>>

Get signature status.

fn get_signature_status_with_commitment( &self, signature: &Signature, commitment_config: CommitmentConfig, ) -> Result<Option<TransactionResult<()>>>

Get signature status. Uses explicit commitment configuration.

fn get_slot(&self) -> Result<u64>

Get last known slot

fn get_slot_with_commitment( &self, commitment_config: CommitmentConfig, ) -> Result<u64>

Get last known slot. Uses explicit commitment configuration.

fn get_transaction_count(&self) -> Result<u64>

Get transaction count

fn get_transaction_count_with_commitment( &self, commitment_config: CommitmentConfig, ) -> Result<u64>

Get transaction count. Uses explicit commitment configuration.

fn get_epoch_info(&self) -> Result<EpochInfo>

fn poll_for_signature_confirmation( &self, signature: &Signature, min_confirmed_blocks: usize, ) -> Result<usize>

Poll until the signature has been confirmed by at least min_confirmed_blocks

fn poll_for_signature(&self, signature: &Signature) -> Result<()>

Poll to confirm a transaction.

fn get_latest_blockhash(&self) -> Result<Hash>

Get last known blockhash

fn get_latest_blockhash_with_commitment( &self, commitment_config: CommitmentConfig, ) -> Result<(Hash, u64)>

Get latest blockhash with last valid block height. Uses explicit commitment configuration.

fn is_blockhash_valid( &self, blockhash: &Hash, commitment: CommitmentConfig, ) -> Result<bool>

Check if the blockhash is valid

fn get_fee_for_message(&self, message: &Message) -> Result<u64>

Calculate the fee for a Message

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.